refactor(icons): centralize icon dictionary, drop ad-hoc switches
- Move the inline `CursorIcon` / `WindsurfIcon` functional components
  out of `ProsePrompt` into `dictionary/icons.ts` so the component no
  longer carries a `@todo`. They stay inline (with full SVG paths)
  until the brand icons land in @bitrix24/b24icons-vue.
- Add the icons that the docs app was importing locally to the same
  dictionary (`DesignIcon`, `FavoriteIcon`, `MoreMIcon`, `NuxtIcon`,
  `AiStarsIcon`, `EncloseTextInCodeTagIcon`, `PlayLIcon`, plus a
  `MdnwebdocsIcon` alias next to the existing `MdnWebDocIcon`), each
  with a one-line note describing its intended usage.
- Replace the per-page `iconFromIconName` helpers in
  `docs/app/pages/index.vue` and `docs/app/pages/docs/[...slug]/index.vue`
  with the central `resolveIcon` from `@bitrix24/b24ui-nuxt/utils`.
  Imports that only fed those switches are dropped; icons used
  directly (`DesignIcon`, `FavoriteIcon`, `MoreMIcon`, `AiStarsIcon`)
  stay as direct imports for readability.
- Sync both icon docs (`icons/{nuxt,vue}.md`) per the file's @memo.
