+++ docs/app/pages/docs/[...slug]/index.vue
@@ -3,13 +3,10 @@ import type { ContentNavigationItem } from '@nuxt/content'
 import { withTrailingSlash, joinURL } from 'ufo' // withoutTrailingSlash
 import { kebabCase } from 'scule'
 import { useColorMode } from '#b24ui/composables/color-mode/useColorMode'
+import { resolveIcon } from '@bitrix24/b24ui-nuxt/utils'
 import DesignIcon from '@bitrix24/b24icons-vue/outline/DesignIcon'
 import FavoriteIcon from '@bitrix24/b24icons-vue/outline/FavoriteIcon'
-import GitHubIcon from '@bitrix24/b24icons-vue/social/GitHubIcon'
 import MoreMIcon from '@bitrix24/b24icons-vue/outline/MoreMIcon'
-import NuxtIcon from '@bitrix24/b24icons-vue/file-type/NuxtIcon'
-import DemonstrationOnIcon from '@bitrix24/b24icons-vue/outline/DemonstrationOnIcon'
-import Bitrix24Icon from '@bitrix24/b24icons-vue/common-service/Bitrix24Icon'
 import AiStarsIcon from '@bitrix24/b24icons-vue/outline/AiStarsIcon'
 
 // const isDev = import.meta.dev
@@ -133,21 +130,6 @@ const communityLinks = computed(() => [
   }
 ])
 
-const iconFromIconName = (iconName?: string) => {
-  if (!iconName) {
-    return undefined
-  }
-
-  switch (iconName) {
-    case 'Bitrix24Icon': return Bitrix24Icon
-    case 'GitHubIcon': return GitHubIcon
-    case 'NuxtIcon': return NuxtIcon
-    case 'DemonstrationOnIcon': return DemonstrationOnIcon
-  }
-
-  return undefined
-}
-
 const explainIcon = computed(() => appConfig.bxAssistant?.icons?.explain || AiStarsIcon)
 
 const showExplainWithAi = computed(() => {
@@ -247,7 +229,7 @@ function makeExplain() {
           :key="link.label"
           :target="link.to?.startsWith('http') ? '_blank' : undefined"
           v-bind="link"
-          :icon="iconFromIconName(link?.iconName)"
+          :icon="resolveIcon(link?.iconName)"
           size="md"
           :b24ui="{ baseLine: 'gap-1.5' }"
         >
