+++ docs/content/docs/1.getting-started/6.integrations/1.icons/2.vue.md
@@ -90,6 +90,32 @@ import GitHubIcon from '@bitrix24/b24icons-vue/social/GitHubIcon'
 import MdnwebdocsIcon from '@bitrix24/b24icons-vue/social/MdnwebdocsIcon'
 import Bitrix24Icon from '@bitrix24/b24icons-vue/common-service/Bitrix24Icon'
 import DemonstrationOnIcon from '@bitrix24/b24icons-vue/outline/DemonstrationOnIcon'
+import DesignIcon from '@bitrix24/b24icons-vue/outline/DesignIcon'
+import FavoriteIcon from '@bitrix24/b24icons-vue/outline/FavoriteIcon'
+import MoreMIcon from '@bitrix24/b24icons-vue/outline/MoreMIcon'
+import NuxtIcon from '@bitrix24/b24icons-vue/file-type/NuxtIcon'
+import AiStarsIcon from '@bitrix24/b24icons-vue/outline/AiStarsIcon'
+import EncloseTextInCodeTagIcon from '@bitrix24/b24icons-vue/editor/EncloseTextInCodeTagIcon'
+import PlayLIcon from '@bitrix24/b24icons-vue/outline/PlayLIcon'
+
+// Inline brand icons — kept here until they ship in @bitrix24/b24icons-vue.
+const CursorIcon: IconComponent = () => h('svg', {
+  'xmlns': 'http://www.w3.org/2000/svg',
+  'viewBox': '0 0 24 24',
+  'aria-hidden': 'true',
+  'data-slot': 'icon'
+}, [
+  h('path', { fill: 'currentColor', d: 'M11.503.131L1.891 5.678…' /* Cursor logo path */ })
+])
+
+const WindsurfIcon: IconComponent = () => h('svg', {
+  'xmlns': 'http://www.w3.org/2000/svg',
+  'viewBox': '0 0 24 24',
+  'aria-hidden': 'true',
+  'data-slot': 'icon'
+}, [
+  h('path', { fill: 'currentColor', d: 'M23.55 5.067a2.177…' /* Windsurf logo path */ })
+])
 
 export default {
   arrowLeft: ArrowToTheLeftIcon,
@@ -132,12 +158,22 @@ export default {
   panelClose: CloseChatIcon,
   panelOpen: OpenChatIcon,
   // Named icons exposed to markdown authors via `iconName`
-  // (ProseCallout, ProsePrompt, ...)
-  GitHubIcon,
-  InfoCircleIcon,
-  MdnWebDocIcon: MdnwebdocsIcon,
-  Bitrix24Icon,
-  DemonstrationOnIcon
+  // (ProseCallout, ProsePrompt, page header / hero CTAs, dropdown actions, ...)
+  GitHubIcon, // GitHub repository / source link
+  NuxtIcon, // Nuxt-specific docs / framework references
+  Bitrix24Icon, // Bitrix24 brand link
+  MdnwebdocsIcon, // MDN documentation references
+  MdnWebDocIcon: MdnwebdocsIcon, // legacy alias used in many ::callout blocks
+  InfoCircleIcon, // generic info / about link
+  DemonstrationOnIcon, // demo / playground / showcase
+  DesignIcon, // "Edit this page" entry point
+  FavoriteIcon, // "Star on GitHub" / favorite CTA
+  MoreMIcon, // dropdown menu trigger
+  AiStarsIcon, // "Explain with AI" / AI assistant entry
+  EncloseTextInCodeTagIcon, // source / "View code" CTA
+  PlayLIcon, // "Try it" / playground / REPL CTA
+  CursorIcon, // Cursor IDE — open prompt in app
+  WindsurfIcon // Windsurf IDE — open prompt in app
 }
 ```
